Implement matrix DWWC
Split into diffusion, matrix & degree_weight modules
